# Conflicts:
#	diffulex/__init__.py
#	diffulex/engine/model_runner.py
#	diffulex_kernel/__init__.py
#	diffulex_kernel/python/dllm_flash_attn_kernels.py
#	test/python/test_linear_fp8.py
#	test/python/test_linear_quantization_module.py
#	test/python/test_quantization_e2e.py
#	test/python/test_quantization_module.py
#	test/python/test_quantization_paths.py
#	test/test_gptq_awq_strategies.py

+++ examples/test_fastdllmv2_diffulex_gsm8k.py
@@ -39,12 +39,13 @@ def summarize_profiling(csv_path: str) -> dict:
 FEW_SHOTS = "<|im_start|>system\nYou are a helpful assistant.<|im_end|>\n"
 
 if __name__ == "__main__":
+    PROFILE = False
     model = "/data1/ckpts/Efficient-Large-Model/Fast_dLLM_v2_7B"
     LLM = Diffulex(
         model,
         use_lora=False,
         model_name="fast_dllm_v2", 
-        enforce_eager=True, 
+        enforce_eager=False, 
         data_parallel_size=1,
         tensor_parallel_size=1,
         gpu_memory_utilization=0.25,
@@ -58,20 +59,21 @@ if __name__ == "__main__":
     tokenizer = AutoTokenizer.from_pretrained(model, trust_remote_code=True)
     sampling_params = SamplingParams(temperature=0.0, max_tokens=256)
     
-    dataset = load_dataset("gsm8k", "main", split="test")["question"][:10]
+    dataset = load_dataset("gsm8k", "main", split="test")["question"][:15]
     prompts = [
         FEW_SHOTS + f"<|im_start|>user\nQuestion: {question}\nAnswer:<|im_end|>\n<|im_start|>assistant\n"
         for question in tqdm(dataset)
     ]
-    
-    output_file = "log/profiles/perf_dvllm_dream_7B.json"
-    if os.path.exists(output_file):
-        os.remove(output_file)
-    # with VizTracer(output_file=output_file, file_info=True) as tracer:
-    #     outputs = llm.generate(prompts[:5], sampling_params)
-    # time.sleep(60)
     s = time.time()
-    outputs = LLM.generate(prompts, sampling_params)
+    if PROFILE:
+        output_file = "log/profiles/perf_dvllm_dream_7B.json"
+        if os.path.exists(output_file):
+            os.remove(output_file)
+        
+        with VizTracer(output_file=output_file, file_info=True) as tracer:
+            outputs = LLM.generate(prompts, sampling_params)
+    else:
+        outputs = LLM.generate(prompts, sampling_params)
     e = time.time()
     print("=*=" * 30, 
           "\nProfiling Results\n",
